// src/app/models/Experiencia.ts
export class Experiencia{
    id?: any;
    position: string;
    company: string;
    img: string;
    mode: string;
    experienceStart: string;
    experienceEnd: string;
    expDescription: string;


// constructor(position:string, company:string, img:string, mode:string, experienceStart:string,
//         experienceEnd:string, expDescription:string) {
    

// this.position = position;
// this.company = company;
// this.img = img;
// this.mode = mode;
// this.experienceStart = experienceStart;
// this.experienceEnd = experienceEnd;
// this.expDescription = expDescription;
// }

}